requirements.txt
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
et-xmlfile==1.1.0
Flask==2.1.3
future==0.18.2
geocoder==1.38.1
gmplot==1.4.1
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gspread==5.4.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Naked==0.1.31
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
sendgrid==6.9.7
shellescape==3.8.1
six==1.16.0
starkbank-ecdsa==2.0.3
twilio==7.11.0
uritemplate==4.1.1
uritemplate.py==3.0.2
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
